+++ arbnode/mel/extraction/abis.go
@@ -1,4 +1,4 @@
-package extractionfunction
+package melextraction
 
 import (
 	"github.com/ethereum/go-ethereum/accounts/abi"
