+++ arbnode/mel/extraction/batch_lookup_test.go
@@ -1,4 +1,4 @@
-package extractionfunction
+package melextraction
 
 import (
 	"context"
