+++ arbnode/mel/extraction/batch_serialization.go
@@ -1,4 +1,4 @@
-package extractionfunction
+package melextraction
 
 import (
 	"context"
