+++ arbnode/mel/extraction/batch_serialization_test.go
@@ -1,4 +1,4 @@
-package extractionfunction
+package melextraction
 
 import (
 	"context"
