+++ arbnode/mel/extraction/messages_in_batch_test.go
@@ -1,4 +1,4 @@
-package extractionfunction
+package melextraction
 
 import (
 	"context"
@@ -12,14 +12,14 @@ import (
 
 	"github.com/offchainlabs/nitro/arbcompress"
 	"github.com/offchainlabs/nitro/arbnode"
-	meltypes "github.com/offchainlabs/nitro/arbnode/message-extraction/types"
+	mel "github.com/offchainlabs/nitro/arbnode/message-extraction/types"
 	"github.com/offchainlabs/nitro/arbos/arbostypes"
 	"github.com/offchainlabs/nitro/arbstate"
 )
 
 func Test_extractMessagesInBatch(t *testing.T) {
 	ctx := context.Background()
-	melState := &meltypes.State{
+	melState := &mel.State{
 		DelayedMessagesRead: 0,
 	}
 	seqMsg := &arbstate.SequencerMessage{
@@ -61,7 +61,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.NoError(t, err)
 		encodedTimestampAdvance, err := rlp.EncodeToBytes(uint64(50))
 		require.NoError(t, err)
-		melState := &meltypes.State{}
+		melState := &mel.State{}
 		seqMsg := &arbstate.SequencerMessage{
 			Segments: [][]byte{
 				append([]byte{arbstate.BatchSegmentKindAdvanceTimestamp}, encodedTimestampAdvance...),
@@ -89,7 +89,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.NoError(t, err)
 		encodedBlockNum, err := rlp.EncodeToBytes(uint64(20))
 		require.NoError(t, err)
-		melState := &meltypes.State{}
+		melState := &mel.State{}
 		seqMsg := &arbstate.SequencerMessage{
 			Segments: [][]byte{
 				append([]byte{arbstate.BatchSegmentKindAdvanceL1BlockNumber}, encodedBlockNum...),
@@ -118,7 +118,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.NoError(t, err)
 		encodedTimestampAdvance := make([]byte, 8)
 		binary.BigEndian.PutUint64(encodedTimestampAdvance, 50)
-		melState := &meltypes.State{}
+		melState := &mel.State{}
 		seqMsg := &arbstate.SequencerMessage{
 			Segments: [][]byte{
 				append([]byte{arbstate.BatchSegmentKindAdvanceTimestamp}, encodedTimestampAdvance...),
@@ -141,7 +141,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.Equal(t, msg.Message.L2msg, []byte("foobar"))
 	})
 	t.Run("delayed message segment greater than what has been read", func(t *testing.T) {
-		melState := &meltypes.State{
+		melState := &mel.State{
 			DelayedMessagesRead: 1,
 		}
 		seqMsg := &arbstate.SequencerMessage{
@@ -164,7 +164,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.Equal(t, arbostypes.InvalidL1Message, msg.Message)
 	})
 	t.Run("gets error fetching delayed message from database", func(t *testing.T) {
-		melState := &meltypes.State{
+		melState := &mel.State{
 			DelayedMessagesRead: 0,
 		}
 		seqMsg := &arbstate.SequencerMessage{
@@ -189,7 +189,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.ErrorContains(t, err, "oops")
 	})
 	t.Run("gets nil delayed message from database", func(t *testing.T) {
-		melState := &meltypes.State{
+		melState := &mel.State{
 			DelayedMessagesRead: 0,
 		}
 		seqMsg := &arbstate.SequencerMessage{
@@ -214,7 +214,7 @@ func Test_extractArbosMessage(t *testing.T) {
 		require.ErrorContains(t, err, "no more delayed messages in db")
 	})
 	t.Run("reading delayed message OK", func(t *testing.T) {
-		melState := &meltypes.State{
+		melState := &mel.State{
 			DelayedMessagesRead: 0,
 		}
 		seqMsg := &arbstate.SequencerMessage{
@@ -256,7 +256,7 @@ func Test_isLastSegment(t *testing.T) {
 		{
 			name: "less than after delayed messages",
 			p: &arbosExtractionParams{
-				melState: &meltypes.State{
+				melState: &mel.State{
 					DelayedMessagesRead: 1,
 				},
 				seqMsg: &arbstate.SequencerMessage{
@@ -270,7 +270,7 @@ func Test_isLastSegment(t *testing.T) {
 		{
 			name: "first segment is zero and next one is brotli message",
 			p: &arbosExtractionParams{
-				melState: &meltypes.State{
+				melState: &mel.State{
 					DelayedMessagesRead: 1,
 				},
 				seqMsg: &arbstate.SequencerMessage{
@@ -284,7 +284,7 @@ func Test_isLastSegment(t *testing.T) {
 		{
 			name: "segment is delayed message kind",
 			p: &arbosExtractionParams{
-				melState: &meltypes.State{
+				melState: &mel.State{
 					DelayedMessagesRead: 1,
 				},
 				seqMsg: &arbstate.SequencerMessage{
@@ -298,7 +298,7 @@ func Test_isLastSegment(t *testing.T) {
 		{
 			name: "is last segment",
 			p: &arbosExtractionParams{
-				melState: &meltypes.State{
+				melState: &mel.State{
 					DelayedMessagesRead: 1,
 				},
 				seqMsg: &arbstate.SequencerMessage{
@@ -329,7 +329,7 @@ type mockDelayedMessageDB struct {
 
 func (m *mockDelayedMessageDB) ReadDelayedMessage(
 	_ context.Context,
-	_ *meltypes.State,
+	_ *mel.State,
 	delayedMsgsRead uint64,
 ) (*arbnode.DelayedInboxMessage, error) {
 	if m.err != nil {
