+++ arbnode/mel/runner/fsm.go
@@ -1,11 +1,11 @@
-package mel
+package melrunner
 
 import (
 	"fmt"
 
 	"github.com/offchainlabs/bold/containers/fsm"
 	"github.com/offchainlabs/nitro/arbnode"
-	meltypes "github.com/offchainlabs/nitro/arbnode/message-extraction/types"
+	"github.com/offchainlabs/nitro/arbnode/mel"
 	"github.com/offchainlabs/nitro/arbos/arbostypes"
 )
 
@@ -46,12 +46,12 @@ type backToStart struct{}
 
 // An action that transitions the FSM to the processing next block state.
 type processNextBlock struct {
-	melState *meltypes.State
+	melState *mel.State
 }
 
 // An action that transitions the FSM to the saving messages state.
 type saveMessages struct {
-	postState       *meltypes.State
+	postState       *mel.State
 	messages        []*arbostypes.MessageWithMetadata
 	delayedMessages []*arbnode.DelayedInboxMessage
 }
